+++ src/miner.cpp
@@ -6,8 +6,6 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <inttypes.h>
-
 #include "miner.h"
 
 #include "core.h"
@@ -17,9 +15,13 @@
 #include "crypto/scrypt.h"
 #include "auxpow.h"
 #include "pow.h"
+#include "util.h"
+#include "utilmoneystr.h"
 #ifdef ENABLE_WALLET
 #include "wallet.h"
 #endif
+
+#include <boost/thread.hpp>
 
 using namespace std;
