+++ src/qt/transactiondesc.cpp
@@ -14,6 +14,7 @@
 #include "transactionrecord.h"
 #include "timedata.h"
 #include "ui_interface.h"
+#include "util.h"
 #include "wallet.h"
 #include "script.h"
