+++ src/qt/utilitydialog.cpp
@@ -22,9 +22,10 @@
 #include "guiutil.h"
 
 #include "init.h"
-#include "util.h"
 #include "net.h"
 #include "version.h"
+
+#include <stdio.h>
 
 #include <QLabel>
 #include <QFont>
