+++ src/rpcserver.cpp
@@ -25,6 +25,7 @@
 #include <boost/iostreams/stream.hpp>
 #include <boost/lexical_cast.hpp>
 #include <boost/shared_ptr.hpp>
+#include <boost/thread.hpp>
 #include "json/json_spirit_writer_template.h"
 
 using namespace boost;
@@ -82,6 +83,11 @@ void RPCTypeCheck(const Object& o,
             throw JSONRPCError(RPC_TYPE_ERROR, err);
         }
     }
+}
+
+static inline int64_t roundint64(double d)
+{
+    return (int64_t)(d > 0 ? d + 0.5 : d - 0.5);
 }
 
 int64_t AmountFromValue(const Value& value)
