+++ src/test/alert_tests.cpp
@@ -12,6 +12,7 @@
 #include "key.h"
 #include "serialize.h"
 #include "util.h"
+#include "utilstrencodings.h"
 #include "version.h"
 
 #include <fstream>
