+++ src/test/allocator_tests.cpp
@@ -4,6 +4,8 @@
 
 #include "util.h"
 
+#include "allocators.h"
+
 #include <boost/test/unit_test.hpp>
 
 BOOST_AUTO_TEST_SUITE(allocator_tests)
