+++ src/test/crypto_tests.cpp
@@ -6,7 +6,7 @@
 #include "crypto/sha1.h"
 #include "crypto/sha2.h"
 #include "random.h"
-#include "util.h"
+#include "utilstrencodings.h"
 
 #include <vector>
